mine

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,17 +10,17 @@
 int main(void)
 {
 	int i = 2;
-	unsigned long long int nums[50];
+	unsigned long int nums[50];
 
 	nums[0] = 1;
 	nums[1] = 2;
 
-	printf("%llu, %llu", nums[0], nums[1]);
+	printf("%lu, %lu", nums[0], nums[1]);
 
 	for (; i < 50; i++)
 	{
 		nums[i] = nums[i - 2] + nums[i - 1];
-		printf(", %llu", nums[i]);
+		printf(", %lu", nums[i]);
 	}
 	printf("\n");
 	return (0);
